feat: 财务邮箱记录。 refactor: 优化用户信息完善逻辑

+++ src/main/java/cn/coal/trading/controller/UserController.java
@@ -229,7 +229,7 @@ public class UserController {
     @PostMapping("/complete")
     public ResponseData complete(@RequestBody CompanyInformation info) {
         TokenProfile profile = ProfileHolder.getProfile();
-        info.setUserId((long) Integer.parseInt(profile.getId()));
+        info.setUserId(Long.parseLong(profile.getId()));
         //  info.setUserId(7L);
         ResponseData result = userService.complete(info);
